[SPARK-45506][CONNECT] Add ivy URI support to SparkConnect addArtifact
### What changes were proposed in this pull request?
This PR extends existing SparkConnect `SparkSession.addArtifact` API to support Apache Ivy URIs in extend to the existing local `.jar` & `.class` file support. It leverages the existing support of Apache Ivy which was previously a part of spark-code (`SparkSubmitUtils`)

This PR contains the following changes:
1. Refactoring `SparkSubmitUtils` and moving Ivy-related parts to `commons` module. This results in adding two small dependencies to `common` (`apache ivy` and `commons-io`)
2. Extracting some parts of `Utils` & `TestUtils` to `common`
3. Extending SparkConnect `SparkSession.addArtifact` to support Ivy URIs

Refactoring changes done in 1-2 are not altering code functionality. These changes are mostly about moving code between files & packages and adjusting some dependencies

### Why are the changes needed?
It's possible to add maven artifacts to Spark, but SparkConnect currently lacks this functionality

### Does this PR introduce _any_ user-facing change?
Yes: it extends the existing SparkConnect `SparkSession.addArtifact` API to support more URI types

### How was this patch tested?
Existing tests covering Ivy resolution functionality;
New test in `ArtifactSuite` testing ivy support in `ArtifactManager`
New E2E test in `ReplE2ESuite` testing adding maven library and using it from UDF

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #43354 from vsevolodstep-db/SPARK-45506/add-artifact-ivy.

Authored-by: Vsevolod Stepanov <vsevolod.stepanov@databricks.com>
