+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/application/ReplE2ESuite.scala
@@ -26,6 +26,8 @@ import org.apache.commons.io.output.ByteArrayOutputStream
 import org.scalatest.BeforeAndAfterEach
 
 import org.apache.spark.sql.test.{IntegrationTestUtils, RemoteSparkSession}
+import org.apache.spark.util.IvyTestUtils
+import org.apache.spark.util.MavenUtils.MavenCoordinate
 
 class ReplE2ESuite extends RemoteSparkSession with BeforeAndAfterEach {
 
@@ -192,6 +194,36 @@ class ReplE2ESuite extends RemoteSparkSession with BeforeAndAfterEach {
     val output = runCommandsInShell(input)
     assertContains("Array[Int] = Array(2, 2, 2, 2, 2)", output)
     // scalastyle:on classforname line.size.limit
+  }
+
+  test("External JAR") {
+    val main = MavenCoordinate("my.great.lib", "mylib", "0.1")
+    IvyTestUtils.withRepository(main, None, None) { repo =>
+      val input =
+        s"""
+           |// this import will fail
+           |import my.great.lib.MyLib
+           |
+           |// making library available in the REPL to compile UDF
+           |import coursierapi.{Credentials, MavenRepository}
+           |interp.repositories() ++= Seq(MavenRepository.of("$repo"))
+           |import $$ivy.`my.great.lib:mylib:0.1`
+           |
+           |val func = udf((a: Int) => {
+           |  import my.great.lib.MyLib
+           |  MyLib.myFunc(a)
+           |})
+           |
+           |// add library to the Executor
+           |spark.addArtifact("ivy://my.great.lib:mylib:0.1?repos=$repo")
+           |
+           |spark.range(5).select(func(col("id"))).as[Int].collect()
+           |""".stripMargin
+      val output = runCommandsInShell(input)
+      // making sure the library was not available before installation
+      assertContains("not found: value my", getCleanString(errorStream))
+      assertContains("Array[Int] = Array(1, 2, 3, 4, 5)", output)
+    }
   }
 
   test("Java UDF") {
