+++ core/pom.xml
@@ -348,16 +348,6 @@
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>org.apache.ivy</groupId>
-      <artifactId>ivy</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>oro</groupId>
-      <!-- oro is needed by ivy, but only listed as an optional dependency, so we include it. -->
-      <artifactId>oro</artifactId>
-      <version>${oro.version}</version>
-    </dependency>
-    <dependency>
       <groupId>org.seleniumhq.selenium</groupId>
       <artifactId>selenium-java</artifactId>
       <scope>test</scope>
